Adding an utility to process a regular image
Processes a regular image and not only an input restricted to the face
